fix 1

+++ tests/models/qwen2_5_omni/test_modeling_qwen2_5_omni.py
@@ -561,7 +561,7 @@ class Qwen2_5OmniModelIntegrationTest(unittest.TestCase):
         text = self.processor.apply_chat_template(self.messages, tokenize=False, add_generation_prompt=True)
         inputs = self.processor(
             text=text, audio=[self.raw_audio], images=[self.raw_image], return_tensors="pt", padding=True
-        )
+        ).to(torch.float16)
 
         expected_input_ids = torch.tensor(
             [
